Refactor indexer tests
