+++ crates/btcio/src/reader/tx_indexer.rs
@@ -78,285 +78,47 @@ impl TxVisitor for ReaderTxVisitorImpl {
 
 #[cfg(test)]
 mod test {
-    use bitcoin::{
-        block::{Header, Version},
-        hashes::Hash,
-        Amount, Block, BlockHash, CompactTarget, ScriptBuf, Transaction, TxMerkleNode,
-    };
-    use strata_l1tx::{filter::indexer::index_block, utils::test_utils::create_tx_filter_config};
-    use strata_primitives::l1::{payload::L1Payload, BitcoinAmount, ProtocolOperation};
-    use strata_test_utils::{
-        bitcoin::{
-            build_test_deposit_request_script, build_test_deposit_script, create_test_deposit_tx,
-            test_taproot_addr,
-        },
-        l2::{gen_params, get_test_signed_checkpoint},
+    use strata_test_utils::tx_indexer::{
+        test_index_deposit_request_with_visitor, test_index_deposit_with_visitor,
+        test_index_multiple_deposits_with_visitor, test_index_no_deposit_with_visitor,
+        test_index_tx_with_multiple_ops_with_visitor,
     };
 
-    use crate::{
-        reader::tx_indexer::ReaderTxVisitorImpl, test_utils::create_checkpoint_envelope_tx,
-    };
-
-    const TEST_ADDR: &str = "bcrt1q6u6qyya3sryhh42lahtnz2m7zuufe7dlt8j0j5";
-
-    //Helper function to create a test block with given transactions
-    fn create_test_block(transactions: Vec<Transaction>) -> Block {
-        let bhash = BlockHash::from_byte_array([0; 32]);
-        Block {
-            header: Header {
-                version: Version::ONE,
-                prev_blockhash: bhash,
-                merkle_root: TxMerkleNode::from_byte_array(*bhash.as_byte_array()),
-                time: 100,
-                bits: CompactTarget::from_consensus(1),
-                nonce: 1,
-            },
-            txdata: transactions,
-        }
-    }
+    use crate::reader::tx_indexer::ReaderTxVisitorImpl;
 
     #[test]
     fn test_index_deposits() {
-        let params = gen_params();
-        let (filter_config, keypair) = create_tx_filter_config(&params);
-        let deposit_config = filter_config.deposit_config.clone();
-        let idx = 0xdeadbeef;
-        let ee_addr = vec![1u8; 20]; // Example EVM address
-        let tapnode_hash = [0u8; 32];
-        let deposit_script =
-            build_test_deposit_script(&deposit_config, idx, ee_addr.clone(), &tapnode_hash);
-
-        let tx = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount),
-            &deposit_config.address.address().script_pubkey(),
-            &deposit_script,
-            &keypair,
-            &tapnode_hash,
-        );
-
-        let block = create_test_block(vec![tx]);
-
-        let tx_entries = index_block(&block, ReaderTxVisitorImpl::new, &filter_config);
-
-        assert_eq!(tx_entries.len(), 1, "Should find one relevant transaction");
-
-        for op in tx_entries[0].contents().protocol_ops() {
-            if let ProtocolOperation::Deposit(deposit_info) = op {
-                assert_eq!(deposit_info.address, ee_addr, "test: dest should match");
-                assert_eq!(deposit_info.deposit_idx, idx, "test: idx should match");
-                assert_eq!(
-                    deposit_info.amt,
-                    BitcoinAmount::from_sat(deposit_config.deposit_amount),
-                    "Deposit amount should match"
-                );
-            } else {
-                panic!("Expected Deposit info");
-            }
-        }
+        let _ = test_index_deposit_with_visitor(ReaderTxVisitorImpl::new, |tx| {
+            tx.contents().protocol_ops().to_vec()
+        });
     }
 
-    // Ignored because deposit request is not included as ops
-    #[ignore]
+    #[ignore = "Ignored because deposit request is not included as ops"]
     #[test]
     fn test_index_txs_deposit_request() {
-        let params = gen_params();
-        let (filter_config, key_pair) = create_tx_filter_config(&params);
-        let mut deposit_config = filter_config.deposit_config.clone();
-
-        let extra_amt = 10000;
-        deposit_config.deposit_amount += extra_amt;
-        let dest_addr = vec![2u8; 20]; // Example EVM address
-        let dummy_block = [0u8; 32]; // Example dummy block
-        let tapnode_hash = [0u8; 32];
-
-        let deposit_request_script = build_test_deposit_request_script(
-            deposit_config.magic_bytes.clone(),
-            dummy_block.to_vec(),
-            dest_addr.clone(),
-        );
-
-        let tx = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount), // Any amount
-            &deposit_config.address.address().script_pubkey(),
-            &deposit_request_script,
-            &key_pair,
-            &tapnode_hash,
-        );
-
-        let block = create_test_block(vec![tx]);
-
-        let tx_entries = index_block(&block, ReaderTxVisitorImpl::new, &filter_config);
-        let dep_reqs = tx_entries
-            .iter()
-            .flat_map(|tx| tx.contents().deposit_reqs())
-            .collect::<Vec<_>>();
-
-        assert_eq!(dep_reqs.len(), 1, "Should find one deposit request");
-
-        for dep_req_info in dep_reqs {
-            assert_eq!(dep_req_info.address, dest_addr, "EE address should match");
-            assert_eq!(
-                dep_req_info.take_back_leaf_hash, dummy_block,
-                "Control block should match"
-            );
-        }
+        let _ = test_index_deposit_request_with_visitor(ReaderTxVisitorImpl::new, |tx| {
+            tx.contents().protocol_ops().to_vec()
+        });
     }
 
     #[test]
     fn test_index_no_deposit() {
-        let params = gen_params();
-        let (filter_config, keypair) = create_tx_filter_config(&params);
-        let deposit_config = filter_config.deposit_config.clone();
-        let tapnode_hash = [0u8; 32];
-        let irrelevant_tx = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount),
-            &test_taproot_addr().address().script_pubkey(),
-            &ScriptBuf::new(),
-            &keypair,
-            &tapnode_hash,
-        );
-
-        let block = create_test_block(vec![irrelevant_tx]);
-
-        let tx_entries = index_block(&block, ReaderTxVisitorImpl::new, &filter_config);
-
-        assert!(
-            tx_entries.is_empty(),
-            "Should not find any relevant transactions"
-        );
+        let _ = test_index_no_deposit_with_visitor(ReaderTxVisitorImpl::new, |tx| {
+            tx.contents().protocol_ops().to_vec()
+        });
     }
 
     #[test]
     fn test_index_multiple_deposits() {
-        let params = gen_params();
-        let (filter_config, keypair) = create_tx_filter_config(&params);
-        let deposit_config = filter_config.deposit_config.clone();
-        let idx1 = 0xdeadbeef;
-        let idx2 = 0x1badb007;
-        let dest_addr1 = vec![3u8; 20];
-        let dest_addr2 = vec![4u8; 20];
-        let tapnode_hash = [0u8; 32];
-
-        let deposit_script1 =
-            build_test_deposit_script(&deposit_config, idx1, dest_addr1.clone(), &tapnode_hash);
-        let deposit_script2 =
-            build_test_deposit_script(&deposit_config, idx2, dest_addr2.clone(), &tapnode_hash);
-
-        let tx1 = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount),
-            &deposit_config.address.address().script_pubkey(),
-            &deposit_script1,
-            &keypair,
-            &tapnode_hash,
-        );
-        let tx2 = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount),
-            &deposit_config.address.address().script_pubkey(),
-            &deposit_script2,
-            &keypair,
-            &tapnode_hash,
-        );
-
-        let block = create_test_block(vec![tx1, tx2]);
-
-        let tx_entries = index_block(&block, ReaderTxVisitorImpl::new, &filter_config);
-
-        assert_eq!(tx_entries.len(), 2, "Should find two relevant transactions");
-
-        for (i, info) in tx_entries
-            .iter()
-            .flat_map(|op_txs| op_txs.contents().protocol_ops())
-            .enumerate()
-        {
-            if let ProtocolOperation::Deposit(deposit_info) = info {
-                assert_eq!(
-                    deposit_info.address,
-                    if i == 0 {
-                        dest_addr1.clone()
-                    } else {
-                        dest_addr2.clone()
-                    },
-                    "test: dest should match for transaction {i}",
-                );
-                assert_eq!(
-                    deposit_info.deposit_idx,
-                    [idx1, idx2][i],
-                    "test: idx should match"
-                );
-                assert_eq!(
-                    deposit_info.amt,
-                    BitcoinAmount::from_sat(deposit_config.deposit_amount),
-                    "test: deposit amount should match for transaction {i}",
-                );
-            } else {
-                panic!("test: expected DepositInfo for transaction {i}");
-            }
-        }
+        let _ = test_index_multiple_deposits_with_visitor(ReaderTxVisitorImpl::new, |op_txs| {
+            op_txs.contents().protocol_ops().to_vec()
+        });
     }
 
-    // TODO; re-enable this when we decide multiple ops can be present with deposits
-    // Also, This could just be okay in the future where protocol Ops is going to be replaced by
-    // some other mechanism.
-    #[ignore]
     #[test]
     fn test_index_tx_with_multiple_ops() {
-        let params = gen_params();
-        let (filter_config, keypair) = create_tx_filter_config(&params);
-        let deposit_config = filter_config.deposit_config.clone();
-        let idx = 0xdeadbeef;
-        let ee_addr = vec![1u8; 20]; // Example EVM address
-        let tapnode_hash = [0u8; 32];
-
-        // Create deposit utxo
-        let deposit_script =
-            build_test_deposit_script(&deposit_config, idx, ee_addr.clone(), &tapnode_hash);
-
-        let mut tx = create_test_deposit_tx(
-            Amount::from_sat(deposit_config.deposit_amount),
-            &deposit_config.address.address().script_pubkey(),
-            &deposit_script,
-            &keypair,
-            &tapnode_hash,
-        );
-
-        // Create envelope tx and copy its input to the above deposit tx
-        let num_envelopes = 1;
-        let l1_payloads: Vec<_> = (0..num_envelopes)
-            .map(|_| {
-                let signed_checkpoint = get_test_signed_checkpoint();
-                L1Payload::new_checkpoint(borsh::to_vec(&signed_checkpoint).unwrap())
-            })
-            .collect();
-        let envtx = create_checkpoint_envelope_tx(&params, TEST_ADDR, l1_payloads);
-        tx.input.push(envtx.input[0].clone());
-
-        // Create a block with single tx that has multiple ops
-        let block = create_test_block(vec![tx]);
-
-        let tx_entries = index_block(&block, ReaderTxVisitorImpl::new, &filter_config);
-        println!("tx_entries: {:?}", tx_entries);
-
-        assert_eq!(
-            tx_entries.len(),
-            1,
-            "Should find one matching transaction entry"
-        );
-        assert_eq!(
-            tx_entries[0].contents().protocol_ops().len(),
-            2,
-            "Should find two protocol ops"
-        );
-
-        let mut dep_count = 0;
-        let mut ckpt_count = 0;
-        for op in tx_entries[0].contents().protocol_ops() {
-            match op {
-                ProtocolOperation::Deposit(_) => dep_count += 1,
-                ProtocolOperation::Checkpoint(_) => ckpt_count += 1,
-                _ => {}
-            }
-        }
-        assert_eq!(dep_count, 1, "should have one deposit");
-        assert_eq!(ckpt_count, 1, "should have one checkpoint");
+        let _ = test_index_tx_with_multiple_ops_with_visitor(ReaderTxVisitorImpl::new, |tx| {
+            tx.contents().protocol_ops().to_vec()
+        });
     }
 }
