+++ crates/test-utils/src/lib.rs
@@ -6,6 +6,7 @@ pub mod bitcoin_mainnet_segment;
 pub mod bridge;
 pub mod evm_ee;
 pub mod l2;
+pub mod tx_indexer;
 
 /// The default buffer size for the `ArbitraryGenerator`.
 const ARB_GEN_LEN: usize = 65_536;
